Fix for issue 431. (i.e. die with better error message if pid file is corrupt)
jeff



git-svn-id: http://anonsvn.internet2.edu/svn/perfSONAR-PS/trunk@4166 752d5dfb-d90f-4296-8230-53451a7f88fe

+++ perfSONAR_PS-perfSONARBUOY/lib/OWP.pm
@@ -94,6 +94,7 @@ sub daemonize {
         $_ = <$fh>;
         if ( defined $_ ) {
             my ( $pid ) = /(\d+)/;
+            die "PID from $args{'PIDFILE'} invalid: $!" unless $pid;
             chomp $pid;
             die "$FindBin::Script:$pid still running..."
                 if ( kill( 0, $pid ) );
